Move tests on retry policy to their own package
